refactor : 뷰 분리 - GameView => InputView(입력 전용 뷰), ResultView(결과 출력 전용 뷰)로 분리
