+++ src/main/java/racinggame/view/InputView.java
@@ -1,13 +1,9 @@
 package racinggame.view;
 
 import nextstep.utils.Console;
-import racinggame.domain.Car;
 import racinggame.utils.ValidationUtils;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class GameView {
+public class InputView {
     private static final String ERROR_MESSAGE = "[ERROR]";
 
     // 자동차 이름 입력받기
@@ -51,24 +47,5 @@ public class GameView {
             return false;
         }
         return true;
-    }
-
-    // 자동차들 위치 출력
-    public void printCars(List<Car> cars) {
-        for (Car car : cars) {
-            String carStatus = String.format("%s : %s", car.getName(), car.getPosition().getPositionToBar());
-            System.out.println(carStatus);
-        }
-        System.out.println();
-    }
-
-    // 승자 출력
-    public static void printWinners(List<Car> winners) {
-        List<String> winnersNames = new ArrayList<>();
-        for (Car winner : winners) {
-            winnersNames.add(winner.getName());
-        }
-        String result = String.join(",", winnersNames);
-        System.out.println(String.format("최종 우승자는 %s 입니다.", result));
     }
 }
